+++ src/app/components/reporting/realtime/realtimeDashboardsController.js
@@ -5,18 +5,15 @@ angular.module('liveopsConfigPanel')
   function($scope, $timeout, dashboard, dashboards, $state, FullscreenService, TableMiddlewareService, _) {
 
     _.forEach(dashboards, function(dash) {
-	    if (!_.isEmpty(dash.activeDashboard)){
-	    	dash.activeDashboard.name = dash.name;
-	    	dash.activeDashboard.id = dash.id;
-	    	dash = dash.activeDashboard;
-	    }
-
       if (dash.id === dashboard) {
-        if (!_.isEmpty(dash.activeDashboard)) {
-          dash.activeDashboard.name = dash.name;
-          dash.activeDashboard.id = dash.id;
+        var currentDashboard = window.allDashboards.find(function (item) {
+          return item.id === dash.id;
+        });
+        if (currentDashboard.dashboardCategory === 'Standard Dashboards') {
+          $scope.dashboard = currentDashboard;
+        } else {
+          $scope.dashboard = currentDashboard.activeDashboard;
         }
-        $scope.dashboard = dash;
       }
     });